+++ app/[locale]/community/faculty-recruitment/page.tsx
@@ -1,4 +1,4 @@
-import { getFacultyRecruitment } from '@/apis/recruitment';
+import { getFacultyRecruitment } from '@/apis/v1/recruit';
 import { EditButton } from '@/components/common/Buttons';
 import LoginVisible from '@/components/common/LoginVisible';
 import HTMLViewer from '@/components/editor/HTMLViewer';
